This is the commit that actually renames basketlistview -> baskettree.
It's a squashed commit of r601 and r602, and has some other changes as well.

As much as rewriting history is a bad thing (tm), it is probably necessary in
this case -- this allows me to preserve the history of the renamed files.

For completeness's sake, the original svn ids are:

git-svn-id: http://basket.linux62.org/svn/basket/branches/kde4port@601 68e0c82f-6715-0410-a4b6-c4571a790354
git-svn-id: http://basket.linux62.org/svn/basket/branches/kde4port@602 68e0c82f-6715-0410-a4b6-c4571a790354

Signed-off-by: Kelvie Wong <kelvie@ieee.org>