Removed all of the whitespace changes from kde4port-svn.
The process to get rid of the whitespace changes was this:
1. First ask git to create a diff without whitespace changes
2. Run a script to go in and reject hunks with non-trivial whitespace
   changes (such as multiline whitespace changes and adding whitespace
   where there were none before).
3. Apply the new patch.
4. Apply the changes between this new HEAD and kde4port-svn onto kde4port-svn

For those interested, step 2 (it's not ???) was done in Emacs using
Emacs lisp calls via Pymacs.

This was some effort, but the rationale behind this is to make it a
_lot_ easier to merge later on.

This was initially done as a squashed merge, but I don't want to get rid of the
history when I merge (credit goes where credit is due).

This saves about 5000 lines to review as I merge this branch into the
fresh port (done by the kdesdk porting scripts) to see what work was
actually done on this port.