Updated the README file with some contributor tips.
There's more to write still (especially on the Git section), but this is
the developer README.  We shouldn't really have any need for a user
README at this point anyways.

It is an asciidoc document, so I'll probably post the HTML somewhere.

I have a feeling this will be split into several documents at some point
in the future...