Ported all of the KAction (and related) functionality.
This includes the old KGlobalAccel calls.

The majority of the changes are in BNPView (as a quick diff --stat should
show).

BNPView::showGlobalShortcutSettingsDialog is also no longer required, as
setting both global and application shortcuts are set in the same dialog
in KDE 4.

An actionCollection() function was also added to MainWindow, as it has
been removed from KMainWindow (for whatever reason, perhaps not every
main window wanted to deal with windows settings).

Again, the large majority of the work on converting the various KActions ctors
were done via a few emacs functions, this would have been hell to do by hand.
It's written in Pymacs/elisp, but I wonder if it's worth sending to the KDE guys
(I'm sure I'm not the only one that has to port KActions from 3 to 4).  They are
a bit far from completely automatic though :)

A TODO for the near future would be to make sure the shortcut dialog works, and
actually includes all of the shortcuts.