Removed StateMenuItem.
It got replaced with a very thin wrapper around KToggleAction.
KToggleAction correctly lays out its checkbox with its icon, so that
renders most of the code in StateMenuItem useless, with two caveats:

There is no longer the radio button display available (which is used when
the emblem icon is pressed in notes), and the background colour is no
longer drawn.  Checkboxes are used in place of the missing radio buttons,
and as for the background colour, that will have to be reconsidered whether
or not we want to do that -- it would involve a custom menu, and is very
likely to make the menu look out of place on other platforms without
some tweaking (which I will argue is better suited for implementation up
higher, in say, KMenu).