Ported all of the kapp->iconLoader() calls.
They're all so ugly.. and long.  I think I'm going to leave the canReturnNull
parameter to false almost always (that's the default, to use the "unknown" icon
when an icon isn't found, rather than return a null pixmap).  It seems the only
place that it really needs to return a null pixmap is in the case of the
function in notefactory.cpp.

Then again... KDE has enough null pixmaps as it is right now.