Ported the keyboard context menu event calls in Basket.
Before, it was using some odd hand-rolled functinos in PopupMenu --
but I really don't see the point.  In the case where there are no notes,
it doesn't matter where the context menu shows up (I just put it at the
top left corner).

In the case where a note is selected (and focussed), we don't need another
class just to pop it up in the middle.  The bottom left of the note
seems to be a good compromise -- it may show above this point if there's
no room at the bottom, but we can't help that.