Porting of the QColors used in the painting done in Basket.
This is just to get it to work initially -- Basket will stay a Q3ScrollView
until the porting is done, and then will get turned into a QGraphicsView
(or perhaps, a scene; the view will display these scenes).  We'll see when
this is all done, but in any case, it will no longer be a widget, and will
no longer have its own palette.