Basket: Re-wrote safelySaveToFile
This is to reflect the new KSaveFile changes -- the comments were also
reworded, and the error messages have changed (I opted to use the default
KSaveFile::errorString's instead).

Also, the delay to the next retry has now changed to ramp up to a maximum
of 2 minutes.  I have a feeling that this may be a bit too long.  Either
way, this should only happen on a rare error anyways.