ported basketproperties.h and basketproperties.cpp to qt4
not sure about the KKeyButton to KShortcutWidget Changes are feature identical, but this is what is recommend in the KDE3 to KDE4 porting guide.
also not sure about what the bQtShortcut do in setShortcut, but I don't think it matters for now. Need to go investigate these nuances in qt4.