Reworked/ported AnimationContent.
I'd like some input on this patch; not quite sure if this 100% what I'd
want.

I think that eventually, this type of note will be deprecated in favour
of something that uses Phonon; that allows both sound and video in many
formats.  Even right now, we don't exactly have Gif support guaranteed.

That said, there are still some other issues with this class -- it reads
the entire file into memory; I think we can do better than that if we set
the IODevice of the movie to a file rather than a QBuffer (which is just a
byte array).  I can imagine a basket taking up a lot of memory with a lot
of animations (I know that Gifs can sometimes be really large).

I also removed the movieStatus function, whose sole function was to check
if it's just an image and not a movie -- I am not quite sure why this is
done; the SVN history doesn't go quite that far.  Either way, this is
something that will require thorough testing and review, I'm sure.

Sebastien left a couple of TODOs here as well; I'm not quite sure why, but
I'll leave them in for now.