Remove Focused* classes in focusedwidgets.
It's been replaced by a very simple event filter implementation.

FocusedTextEdit will have to go next, as it seemed to have become the
dumping ground for the TextEdit-related stuff.  Although, currently, it can
be replaced by a simple event filter, I think there is good reason to keep
our own subclass of KTextEdit around, in case we need to do anything with
it in the future.  It just probably does not belong in this file, with
this name.

Also, my original intention was to have a single event filter for each
widget, but I quickly broke that rule in InlineEditors -- it doesn't
matter there, though.