FocusedTextEdit: Removed the createPopupMenu function.
This function did two things: disable spellcheck and disable the "Allow
Tabulations" item.

The spell check should be left in -- that's KDE's job to make sure it
works.  The latter thing (finding the actions in a menu) is no longer
possible; the default is to have "Allow Tabulations" on anyways, so
hopefully we won't affect anything.

I do plan on adding this back when I can have access to the menu items (I
put it on the TODO wiki), but this functionality will have to go in
kdelibs.