FocusedTextEdit: Rewrote the key event handler.
There was a lot of cruft in here that was removed; I didn't like the remapping
of the <Return> and <C-Return> keys; rather than try to be consistent with old
modes, we should rather strive to be consistent with KDE standards (which make
more sense across platforms).

Calling updatesEnabled also implicitly updates the widget and its children, so
we don't need to call updateContents (which doesn't exist anymore).