SystemTray: Put away most of the custom SystemTray code for now.
KSystemTrayIcon2 has now been moved out into its own file (and hopefully
will amend onto the existing KSystemTrayIcon into the future).

As for our SystemTray class; the unported stuff has been placed inside
an #ifdef; as parts of it get ported, they will be moved up into the
"clean" SystemTray class; the old one is named SystemTray2, and is not
compiled (currently).

So far, the only things implemented for SystemTray is the icon and tooltip
handling; as things are added to SystemTray, no changes to other code
_should_ need to be handled.