Replaced KColorCombo2 with a thin wrapper class around KColorCombo.
This is until we actually do port KColorCombo2 (which we _need_ to, it was
so much better).  When we do, it should adhere to KColorCombo's naming
conventions, though (such as using the activated() and hovered() signals
rather than changed()).

The old one is left in the code with an #ifdef.