Fix the rich text notes so that they don't save on every keystroke.
There is a bug in current 1.0.2 version that causes me a lot of problem, so that
I had to patch it manually. While writing a note, every single keystroke causes
a file to be rewritten to the disk. This is causes a huge performance hit while
writing notes (just hold down any key and watch your HDD led blink
continuously). This makes basket pretty unusable if your $HOME directory is
mounted over NFS.

I had to dig basket source code and fix it myself in order to use basket in my
environment. I haven't studied basket source code thoroughly, and I don't know
if the fix is really correct, but now it doesn't causes disk writes for each
keystroke anymore, making it possible to use it over NFS without huge
performance impact.

There is still one situation where intense disk writes happen: when the note
area is being resized to fit more text (either in width in free layout mode, or
in height when more lines are added). So, the patch partially fixes the
problem. It is still necessary to find where this other disk access happen and
fix it also.

Addenum (kw):
This was fixed in the (now-obsolete) Text notes, but the fix didn't see its way
into the rich text note types, which replace it.

Message-edited-by: Kelvie Wong <kelvie@ieee.org>