Ported the tooltips in BasketListView.
Hopefully, the tooltips will pop up in the right spot on every platform; it says
in the Qt docs that the spot that the tooltip will show up on differs from
platform to platform.  In this case, we want the tooltip to completely cover the
list view item (for longer names).  This may require tweaking in the future.