Don't use setCornerWidget(0) on a Q3ScrollView.
The docs say it's OK, but here's the implementation:

void Q3ScrollView::setCornerWidget(QWidget* corner)
{
    QWidget* oldcorner = d->corner;
    if (oldcorner != corner) {
        if (oldcorner) oldcorner->hide();
        d->corner = corner;
        corner->setParent(this);
        updateScrollBars();
        if (corner) corner->show();
    }
}

Yes.  It dereferences it to set the parent without checking for NULL.  The
bug report has also been sent to Trolltech, but I have a feeling that this
will be ported to use QGraphicsView by the time this is fixed and released.