Fixed the tag menu to actually work.
This was all due to what seems like the (obsolete) QMenu::activated(int) signal
not being emitted at all.  In turn, I abused the QAction::data property to pass
in the integer id's that are now deprecated in the QMenu and related classes.

I'm not particularly fond of this implementation, and I wish I could read into
the Trolls' minds to see what they had in mind to replace the "action ID" way of
handling many actions.  Maybe we can split up toggledTagInMenu into multiple
functions in the future?