Fix Password Dialog (bug 0000034) and update to Qt4
The Password Dialog that you obtain by right clicking on a basket in the tree
view and then selecting "Password" wasn't displaying anything because the
password widget was set to be a child of the current main widget, rather than
of the password dialog.  Also, the password widget used the deprecated
Q3ButtonGroup and Q3GroupBox classes.  I updated the .ui file to use
QButtonGroup instead, which required some changes to the "type" and "setType"
functions, then I fixed the PasswordDlg constructor so that the password widget
showed up.  This necessitated changing the password widget to inherit from
QWidget rather than KDialog so that we didn't get spurious OK and Cancel
buttons behind the widget.