Restore the auto-hide functionality in the FilterBar.
Now the FilterBar behaves a lot better -- I think this is what Sébastien
had in mind for the filter bar originally (it is apparent in some of the
code that is commented out and the comments themselves).

It now checks the editor focus before hiding; FilterBar:;reset() now closes
the filter bar.

With this new change, the "Filter all Baskets" action now automatically
shows the filter bar initially.