Fix exporting and importing of Basket Archives.  Resolves issue 0000039.
This resolves several issues with exporting and importing basket archives.  Bug
0000039 was caused by code in archive.cpp that mixed stream operations with the
file.write() function.  Because QTextStream buffers data, this was causing the
file header (with version information, etc) to not be written to the file in
the appropriate place.  By flushing the stream before using file.write() in two
places, I was able to resolve this bug.

Unfortunately, another problem -- an infinite loop in FormatImporter -- made
importing impractical.  What happens is that if the importer tries to read an
icon that wasn't saved in the archive, it loops forever waiting for that file
to "finish copying".  Of course, this is impossible and the program freezes.

I am not sure under what circumstances that corner case can appear, but
checking for the existence of the file before trying to copy it seems to make
importing work.  I was able to export and import baskets.