Fix KAction bug which caused a crash.
When right clicking on an emblem and selecting "customize" the dialog
would sometimes crash.  This happened only when the dialog had multiple
states.  The first time you selected "customize" nothing would happen
but the dialog would fail to appear.  The second time, the program would
crash.  This was caused by mixing and matching of old Qt3
KMenu::insertItem calls with the new Qt4 QAction class.  I have modified
the code to use only the new system.