Convert bnpview to a QTreeWidget to fix dragging.
Coververted the BasketListView class from a Q3ListView to a QTreeWidget.
It probably needs extensive testing, though!  There are some significant
differences in how a QTreeWidget works internally.  For instance, it is no longer
possible to iterate through the items as if they were a linked list.
Instead, you can use a for loop to access each child of a node.

There are also some very serious changes in how the QTreeWidget
interacts with other classes.  The names of most of the signals have
changed and/or are no longer available.  I removed a lot of partially
implemented or unimplemented code to make this work.  It can easily be
put back in if necessary.