Reworked our DatabaseObject class
It is no longer a pure virtual, as the plan is to have copies of this
class passed back and forth.

The only pressing issue with this new design is the data type of the raw
data returned by data().  Ideally, data() would return a QIOStream, so we
can accommodate other forms of reading (such as buffered reading) rather
than read all of the data into memory at once into a QByteArray, but then
the task of making sure the memory is freed later is left to the caller,
which I do want to avoid.

The new implementation also uses pimpl, which should hopefully make this
more BC friendly in the future.