First implementation of the BasketDatabase.
No, it has not been tested yet, it will have to get refined as the client
code is moved toward this new system.

This database uses the SHA1 hash as the key into the database, and
requires three directories in the user's data directory to store all of
its information.  It also uses reference counting to garbage collect.

The only sketchy part of this that I feel is the QDom* calls.  I'm not
particularly a big fan of XML or DOM, but it seemed like the simplest
way to store the object files, at the time without writing my own parser.