Save baskets and notes using the database.
Right now, the mechanism works in parallel with the old mechanism; that is
it is still safe in that it doesn't break existing functionality.  This will
happen when the loading mechanism is put into place, however.

BNPView::save() will now have to be called every single time anything is
changed as it is in charge of changing the database's root object, and that
object depends on the contents of all of its children item(s).