Fix handle selection bug.
I have no idea why this code is here.  There MUST be some reason, right?
However, it is causing a bug when the user selects several notes by left
clicking on the handle using CTRL or SHIFT.  The selection works fine,
but faster than the eye can see, the selection is automaticaly removed
as soon as the mouse button is released.  :-P