Fix a memory access bug.
If you had the editor open when closing basket, you got a crash, because
the destructor tried to save the note, but the note had already been
deleted by a previous destructor....  Not good.  Adding an explicit
delete of m_baskets fixes this issue and potentially fixes several
memory leaks, too!