Cross References are now optional at the note/tag level.
This patch allows the user to change if a note is allowed to try to convert a [[text]] link into a cross reference
by using a checkbox on in the tags dialog. It sets the default value on the Code tag to false if there isn't a tags.xml file.